autoregister plugin API
